feat add Coder

+++ tcp/Message/Message.go
@@ -4,38 +4,38 @@ import "encoding/json"
 
 type Message struct {
 	t    MsgType
-	key  uint64
-	len  int
+	key  uint32
+	len  uint32
 	Data []byte
 }
 
-func NewMessage(t MsgType, key uint64, data []byte) *Message {
+func NewMessage(t MsgType, key uint32, data []byte) *Message {
 	return &Message{
 		t:    t,
 		key:  key,
-		len:  len(data),
+		len:  uint32(len(data)),
 		Data: nil,
 	}
 }
-func AckMessage(key uint64) *Message {
+func AckMessage(key uint32) *Message {
 	return NewMessage(ACK, key, nil)
 }
-func HeartMessage(key uint64) *Message {
+func HeartMessage(key uint32) *Message {
 	return NewMessage(HeartBeat, key, nil)
 }
-func JsonMessage(key uint64, target any) (*Message, error) {
+func JsonMessage(key uint32, target any) (*Message, error) {
 	data, err := json.Marshal(target)
 	if err != nil {
 		return nil, err
 	}
 	return NewMessage(Json, key, data), nil
 }
-func TextMessage(key uint64, text string) *Message {
+func TextMessage(key uint32, text string) *Message {
 	return NewMessage(Text, key, []byte(text))
 }
-func BlobMessage(key uint64, blob []byte) *Message {
+func BlobMessage(key uint32, blob []byte) *Message {
 	return NewMessage(Blob, key, blob)
 }
-func (m *Message) Len() int {
+func (m *Message) Len() uint32 {
 	return m.len
 }
